+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Introduce `icon` configuration field in `MatrixMessengerBaseConfiguration`
 - SDK,UI: Add new search API to include search providers (e.g., an LDAP search) to look for when searching users (see README.md for more details); there is a _feature flag_ `enableNewSearch` which can be set to `false` to use the old search logic and UI
+- SDK,UI: new search: errors of search providers are collected and displayed above search bar
 - SDK: Added check for downloadsDisabled before downloading file
 
 ### Changed
 
+- SDK: Return error in search when inviting user ID with invalid homeserver
 - DEPENDENCY: update Kotlin to 2.3.21
 - DEPENDENCY: update Lognity to 2.2.0
 - DEPENDENCY: update Trixnity to 5.7.0-SNAPSHOT.CI-20260623.160101
@@ -24,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - INTERNAL: pinned the ktfmt IntelliJ plugin to 1.3.0.63
 - UI, SDK: Display and redact reactions still in the outbox
 - SDK: (**breaking change**) Changed `GetEventReactions.EventReaction` to use `EventIdOrTransactionId` instead of `EventId` in order to support reactions in outbox
+- DEPENDENCY: update compose to 1.11.1
 
 ### Deprecated